+++ andes/models/renewable/__init__.py
@@ -4,7 +4,7 @@ Top-level package for renewable models.
 
 from andes.models.renewable.regca1 import REGCA1      # NOQA
 from andes.models.renewable.reeca1 import REECA1      # NOQA
-from andes.models.renewable.reeca1e import REECA1E    # NOQA
+from andes.models.renewable.reeca1e import REECA1E, REECA1G    # NOQA
 from andes.models.renewable.repca1 import REPCA1      # NOQA
 from andes.models.renewable.wtara1 import WTARA1      # NOQA
 from andes.models.renewable.wtarv1 import WTARV1      # NOQA