+++ andes/models/renewable/reeca1e.py
@@ -73,3 +73,55 @@ class REECA1E(REECA1EData, REECA1EModel):
     def __init__(self, system, config):
         REECA1EData.__init__(self)
         REECA1EModel.__init__(self, system, config)
+
+
+class REECA1GData(REECA1Data):
+    """
+    Data for REECA1G.
+    """
+
+    def __init__(self):
+        REECA1Data.__init__(self)
+        self.Kf = NumParam(default=0.0,
+                           info='gain for frequency deviation',
+                           tex_name='K_{df}',
+                           )
+
+        self.sg = IdxParam(info='synchronous gen idx',
+                           model='Synchronous',
+                           default=None,
+                           mandatory=True,
+                           )
+
+
+class REECA1GModel(REECA1Model):
+    """
+    Model for REECA1G. See docstring for REECA1G.
+    """
+
+    def __init__(self, system, config):
+        REECA1Model.__init__(self, system, config)
+
+        self.omega = ExtAlgeb(model='SynGen',
+                              src='omega',
+                              indexer=self.sg,
+                              export=False,
+                              info='generator speed',
+                              unit='pu',
+                              )
+
+        self.Pref.e_str += '- Kf * (omega - 1)'
+
+
+class REECA1G(REECA1GData, REECA1GModel):
+    """
+    REECA1G is a variant of REECA1E.
+
+    REECA1G uses speed from synchronous generators.
+
+    The application of this model is limited because it is uncommon
+    to connect a SynGen on the same bus as a RenGen.
+    """
+    def __init__(self, system, config):
+        REECA1GData.__init__(self)
+        REECA1GModel.__init__(self, system, config)